chore: added the recognition to the home page
