+++ frontend/asl4all/src/App.js
@@ -1,13 +1,12 @@
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom'
 import Home from './pages/home'
-import Test from "./pages/test";
 
 export default function App() {
     return (
         <Router>
             <Routes>
                 <Route path='/ASL4All' element={<Home/>} />
-                <Route path='/test' element={<Test/>} />
+                <Route path="/*" element={<Navigate to="/asl4all" />} />
             </Routes>
         </Router>
     )
