+++ frontend/asl4all/src/pages/home.module.css
@@ -32,6 +32,8 @@
     
     backdrop-filter: blur(5px);
     background-color: rgba(255, 255, 255, 0.75);
+
+
 }
 
 .text_box > div {
